fixing incorrectly labeled test file
